Conflicts:
	src/backend.c
	src/backend.h
	src/connection.c
	src/listener.c
	src/listener.h
	src/table.c
	src/table.h
	tests/Makefile.am

+++ src/table.h
@@ -50,15 +50,14 @@ int accept_table_arg(struct Table *, char *);
 void add_table(struct Table_head *, struct Table *);
 struct Table *table_lookup(const struct Table_head *, const char *);
 const struct Address *table_lookup_server_address(const struct Table *,
-        const char *, unsigned);
+                                                  const char *, size_t);
 void print_table_config(FILE *, struct Table *);
 int valid_table(struct Table *);
 void free_table(struct Table *);
 void init_table(struct Table *);
 
 static inline struct Backend *
-table_lookup_backend(const struct Table *table, const char *name, unsigned name_size)
-{
+table_lookup_backend(const struct Table *table, const char *name, size_t name_size) {
     return lookup_backend(&table->backends, name, name_size);
 }
